+++ CHANGELOG.md
@@ -7,10 +7,11 @@
 
 ### TODO
 
+* [92](https://github.com/b3log/vditor/issues/92) 现在toolbar里面有演示的按钮吗 `question`
 * [89](https://github.com/b3log/vditor/issues/89) 自定义 toolbar 事件、添加按钮到 toolbar 上 `feature`
 * [86](https://github.com/b3log/vditor/issues/86) 代码分包优化 `feature`
 
-### v1.7.1 / 2019-08-07
+### v1.7.2 / 2019-08-07
 
 * [93](https://github.com/b3log/vditor/issues/93) 新增预览模式设置接口 `feature`
 * [91](https://github.com/b3log/vditor/issues/91) No default jsDelivr CDN file set `development`